+++ Src/Apps/krull/SqliteBackEnd.h
@@ -30,7 +30,7 @@ public:
 protected:
 	bool				Execute				(const string& sql);
 	string				FromInt				(int value) const;
-	string				FromUInt			(unsigned int value) const;
+	string				FromUInt			(size_t value) const;
 	string				FromFloat			(float value) const;
 
 private:
